+++ lib/input/tcp.go
@@ -1,0 +1,108 @@
+// Copyright (c) 2019 Ashley Jeffs
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package input
+
+import (
+	"io"
+	"net"
+
+	"github.com/Jeffail/benthos/lib/input/reader"
+	"github.com/Jeffail/benthos/lib/log"
+	"github.com/Jeffail/benthos/lib/metrics"
+	"github.com/Jeffail/benthos/lib/types"
+)
+
+//------------------------------------------------------------------------------
+
+func init() {
+	Constructors[TypeTCP] = TypeSpec{
+		constructor: NewTCP,
+		description: `
+Connects to a TCP server and consumes a continuous stream of messages.
+
+If multipart is set to false each line of data is read as a separate message. If
+multipart is set to true each line is read as a message part, and an empty line
+indicates the end of a message.
+
+If the delimiter field is left empty then line feed (\n) is used.`,
+	}
+}
+
+//------------------------------------------------------------------------------
+
+// TCPConfig contains configuration values for the TCP input type.
+type TCPConfig struct {
+	Address   string `json:"address" yaml:"address"`
+	Multipart bool   `json:"multipart" yaml:"multipart"`
+	MaxBuffer int    `json:"max_buffer" yaml:"max_buffer"`
+	Delim     string `json:"delimiter" yaml:"delimiter"`
+}
+
+// NewTCPConfig creates a new TCPConfig with default values.
+func NewTCPConfig() TCPConfig {
+	return TCPConfig{
+		Address:   "localhost:4194",
+		Multipart: false,
+		MaxBuffer: 1000000,
+		Delim:     "",
+	}
+}
+
+//------------------------------------------------------------------------------
+
+// NewTCP creates a new TCP input type.
+func NewTCP(conf Config, mgr types.Manager, log log.Modular, stats metrics.Type) (Type, error) {
+	delim := conf.TCP.Delim
+	if len(delim) == 0 {
+		delim = "\n"
+	}
+	var conn net.Conn
+	rdr, err := reader.NewLines(
+		func() (io.Reader, error) {
+			if conn != nil {
+				conn.Close()
+				conn = nil
+			}
+			var err error
+			conn, err = net.Dial("tcp", conf.TCP.Address)
+			return conn, err
+		},
+		func() {
+			if conn != nil {
+				conn.Close()
+				conn = nil
+			}
+		},
+		reader.OptLinesSetDelimiter(delim),
+		reader.OptLinesSetMaxBuffer(conf.TCP.MaxBuffer),
+		reader.OptLinesSetMultipart(conf.TCP.Multipart),
+	)
+	if err != nil {
+		return nil, err
+	}
+	return NewReader(
+		TypeTCP,
+		reader.NewPreserver(rdr),
+		log, stats,
+	)
+}
+
+//------------------------------------------------------------------------------